mobile flutter  make view for permission

+++ final_project_backend/digital_livestock/signals.py
@@ -1,3 +1,8 @@
+
+
+
+
+ 
 from django.db.models.signals import  post_save,pre_save,post_delete,m2m_changed
 from django.dispatch import receiver
 from django.contrib.auth.models import Group
@@ -69,15 +74,14 @@ def post_save_farm(sender, instance:farm,raw,created:bool, *args, **kwargs):
 					}],
 					'connection_timeout_seconds': 2
 			})
-    print(created)
-    if (instance.old_kwargs["id"]!="" and instance.id!=instance.old_kwargs["id"] and not  created):
+    if (getattr(instance,'old_kwargs',None)and instance.old_kwargs["id"]!="" and instance.id!=instance.old_kwargs["id"] and not  created):
         try:
             client.collections['farm'].documents[instance.old_kwargs["id"]].delete( )
         except Exception as e:
             pass
         client.collections['farm'].documents.import_(json.dumps({"id":instance.id,"name":instance.farm_name}))
         return
-    elif((instance.farm_name!=instance.old_kwargs['farm_name'] )and not  created)  :
+    elif(getattr(instance,'old_kwargs',None)and (instance.farm_name!=instance.old_kwargs['farm_name'] )and not  created)  :
         client.collections['farm'].documents[instance.id].update(json.dumps({"id":instance.id,"name":instance.farm_name }),{'dirty_values': 'coerce_or_reject'})
     else:
          client.collections['farm'].documents.import_(json.dumps({"id":instance.id,"name":instance.farm_name}))
@@ -89,14 +93,12 @@ def post_save_farm(sender, instance:farm,raw,created:bool, *args, **kwargs):
                         os.remove(instance.old_kwargs["img"].path)
     except Exception as e:
         pass
-        print(e)
     m=instance.__dict__.copy()
     m.pop("old_kwargs")
     instance.old_kwargs=m
     instance.old_kwargs["img"]=ImageFieldFile(instance=instance,field= models.ImageField(upload_to="farm_img",blank=False,null=False),name=instance.img.name)
 @receiver(m2m_changed, sender=User.groups.through, dispatch_uid= 'my_unique_identifier')
 def m2m_group_User(sender, instance:User, action, *args, **kwargs):
-    print(action)
     client = typesense.Client({
 					'api_key': 'AA3jvgcuaEfuB3GAtWjNS3LG66404bd6KHOBK1YqstLgBTtT',
 					'nodes': [{
@@ -113,7 +115,6 @@ def m2m_group_User(sender, instance:User, action, *args, **kwargs):
     elif(action in["post_add"]):
      
       if(Group.objects.all().get(name="farmer") in instance.old_kwargs["group"] and Group.objects.all().get(name="farmer")  not in instance.groups.all() ):
-        print("kwargs"*12)
         client.collections['farmer'].documents[instance.ssn].delete( )
       if(Group.objects.all().get(name="farmer") not in instance.old_kwargs["group"] and Group.objects.all().get(name="farmer")    in instance.groups.all() ):
         client.collections['farmer'].documents.import_(json.dumps({"id":instance.ssn,"name":instance.fname+" "+instance.lname}))
@@ -133,37 +134,34 @@ def post_save_User(sender, instance:User,raw,created:bool, *args, **kwargs):
 					}],
 					'connection_timeout_seconds': 2
 			})
-    if (instance.ssn!=instance.old_kwargs["ssn"] and not  created):
+    if (getattr(instance,'old_kwargs',None)and instance.ssn!=instance.old_kwargs["ssn"] and not  created):
         client.collections['farmer'].documents[instance.old_kwargs["id"]].delete( )
         client.collections['farmer'].documents.import_(json.dumps({"id":instance.ssn,"name":instance.fname+" "+instance.lname}))
         return
-    elif((instance.fname!=instance.old_kwargs['fname'] or instance.lname!=instance.old_kwargs['lname'] )and not  created)  :
+    elif(getattr(instance,'old_kwargs',None)and (instance.fname!=instance.old_kwargs['fname'] or instance.lname!=instance.old_kwargs['lname'] )and not  created)  :
         client.collections['farmer'].documents[instance.ssn].update(json.dumps({"id":instance.ssn,"name":instance.fname+" "+instance.lname}),{'dirty_values': 'coerce_or_reject'})
     else:
         client.collections['farmer'].documents.import_(json.dumps({"id":instance.ssn,"name":instance.fname+" "+instance.lname}))
     try:
         
         if(instance.img.path!=instance.old_kwargs["img"].path):
-            print(instance.img.path)
-            print(instance.old_kwargs["img"].path)
             if os.path.exists(instance.old_kwargs["img"].path):
                         os.remove(instance.old_kwargs["img"].path)
     except Exception as e:
         pass
-        print(e)
     m=instance.__dict__.copy()
     m.pop("old_kwargs")
     instance.old_kwargs=m
     instance.old_kwargs["img"]=ImageFieldFile(instance=instance,field=instance.img.field,name=instance.img.name)
     instance.old_kwargs["group"]=list(instance.groups.all())
-    """
+ 
     if(created):
         totpy1=totpyUsers()
         totpy1.user=instance
         str1=pyotp.random_base32()
         totpy1.totp=str1
         totpy1.save()
-        """
+   
 @receiver(post_delete, sender=User, dispatch_uid="my_unique_identifier")
 def  post_delete_User(sender, instance:User , *args, **kwargs):
     try:
